Shining new and polished artisan commands.

+++ app/Console/Commands/TenantMigrate.php
@@ -5,12 +5,13 @@ namespace App\Console\Commands;
 use Illuminate\Console\Command;
 use App\Services\TenantManager;
 use App\Models\Tenant;
+use Symfony\Component\Console\Exception\RuntimeException;
 
 class TenantMigrate extends Command
 {
-    protected $signature = 'tenants:migrate';
+    protected $signature = 'tenant:migrate {tenantId?}';
 
-    protected $description = 'Run tenant migrations for all registered tenants';
+    protected $description = 'Run tenant migrations for provided tenant or for all registered tenants';
 
     protected $tenantManager;
 
@@ -26,13 +27,34 @@ class TenantMigrate extends Command
 
     public function handle()
     {
-        $tenants = Tenant::all();
+        $tenantId = $this->argument('tenantId');
 
-        foreach ($tenants as $tenant) {
+        if ($tenantId) {
+            $tenant = Tenant::find($tenantId);
+            if (!$tenant) {
+                throw new RuntimeException('Tenant with ID = ' . $tenantId . ' does not exist.');
+            }
+
             $this->tenantManager->setTenant($tenant);
             \DB::purge('tenant');
             $this->migrate();
+
+            return;
         }
+
+        if (!$tenantId && $this->confirm('Do you wish to run migration for ALL tenants?')) {
+
+            $tenants = Tenant::all();
+
+            foreach ($tenants as $tenant) {
+                $this->tenantManager->setTenant($tenant);
+                \DB::purge('tenant');
+                $this->migrate();
+            }
+
+            return;
+        }
+
     }
 
     private function migrate()
